moved various image IO functions to kubric.utils
