+++ kubric/utils.py
@@ -27,21 +27,34 @@
 # limitations under the License.
 import argparse
 import copy
+import functools
+import io
 import json
 import logging
+import multiprocessing.pool
 import pathlib
 import pickle
 import pprint
 import shutil
 import sys
 import tempfile
+import threading
+from typing import Dict, Tuple
 
+import imageio
 import numpy as np
+import png
 import tensorflow as tf
+from tensorflow_datasets.core.utils.type_utils import PathLike
 import tensorflow_datasets.public_api as tfds
+
 
 logger = logging.getLogger(__name__)
 
+
+# --------------------------------------------------------------------------------------------------
+# Kubric argparser
+# --------------------------------------------------------------------------------------------------
 
 class ArgumentParser(argparse.ArgumentParser):
   def __init__(self, *args, **kwargs):
@@ -80,35 +93,9 @@ class ArgumentParser(argparse.ArgumentParser):
       flags = super(ArgumentParser, self).parse_args(args=args)
     return flags
 
-def str2path(strpath):
-  return tfds.core.as_path(strpath)
-
-def setup_directories(FLAGS):
-  assert FLAGS.scratch_dir is not None
-  scratch_dir = FLAGS.scratch_dir
-  if scratch_dir.exists():
-    logging.info("Deleting content of old scratch-dir: %s", scratch_dir)
-    shutil.rmtree(scratch_dir)
-  scratch_dir.mkdir(parents=True)
-  logging.info("Using scratch directory: %s", scratch_dir)
-
-  output_dir = tfds.core.as_path(FLAGS.job_dir)
-  output_dir.mkdir(parents=True, exist_ok=True)
-  logging.info("Using output directory: %s", output_dir)
-  return scratch_dir, output_dir
-
-
-def is_local_path(path):
-  """ Determine if a given path is local or remote. """
-  first_part = pathlib.Path(path).parts[0]
-  if first_part.endswith(':') and len(first_part) > 2:
-    return False
-  else:
-    return True
-
 
 # --------------------------------------------------------------------------------------------------
-# --------------------------------------------------------------------------------------------------
+# Helpers for workers
 # --------------------------------------------------------------------------------------------------
 
 def setup_logging(logging_level):
@@ -119,10 +106,22 @@ def log_my_flags(flags):
   flags_string = pprint.pformat(vars(flags), indent=2, width=100)
   logger.info(flags_string)
 
+
+def done():
+  logging.info("Done!")
+
+  # -- report generated_images to hyperparameter tuner
+  import hypertune
+
+  hpt = hypertune.HyperTune()
+  hpt.report_hyperparameter_tuning_metric(
+      hyperparameter_metric_tag="answer",
+      metric_value=42)
+
+
 # --------------------------------------------------------------------------------------------------
 # Collect metadata
 # --------------------------------------------------------------------------------------------------
-
 
 def get_scene_metadata(scene, **kwargs):
   metadata = {
@@ -186,6 +185,38 @@ def process_collisions(collisions, scene):
   } for c in collisions]
 
 
+# --------------------------------------------------------------------------------------------------
+# File IO helpers
+# --------------------------------------------------------------------------------------------------
+
+def str2path(strpath):
+  return tfds.core.as_path(strpath)
+
+
+def setup_directories(FLAGS):
+  assert FLAGS.scratch_dir is not None
+  scratch_dir = FLAGS.scratch_dir
+  if scratch_dir.exists():
+    logging.info("Deleting content of old scratch-dir: %s", scratch_dir)
+    shutil.rmtree(scratch_dir)
+  scratch_dir.mkdir(parents=True)
+  logging.info("Using scratch directory: %s", scratch_dir)
+
+  output_dir = tfds.core.as_path(FLAGS.job_dir)
+  output_dir.mkdir(parents=True, exist_ok=True)
+  logging.info("Using output directory: %s", output_dir)
+  return scratch_dir, output_dir
+
+
+def is_local_path(path):
+  """ Determine if a given path is local or remote. """
+  first_part = pathlib.Path(path).parts[0]
+  if first_part.endswith(':') and len(first_part) > 2:
+    return False
+  else:
+    return True
+
+
 def save_as_pkl(filename, data):
   with tf.io.gfile.GFile(filename, "wb") as fp:
     logging.info(f"Writing to {fp.name}")
@@ -205,13 +236,129 @@ def save_as_json(filename, data):
     json.dump(data, fp, sort_keys=True, indent=4, cls=NumpyEncoder)
 
 
-def done():
-  logging.info("Done!")
+def write_scaled_png(data: np.array, path_prefix: PathLike) -> Dict[str, float]:
+  """Writes data as pngs to path and optionally returns rescaling values.
 
-  # -- report generated_images to hyperparameter tuner
-  import hypertune
+  data is expected to be of shape [B, H, W, C] and has to have
+  between 1 and 4 channels. A two channel image is padded using a zero channel.
+  uint8 and uint16 are written as is.
+  other dtypes are rescaled to uint16. When rescaling, this method returns the
+  min and max values of the original data.
 
-  hpt = hypertune.HyperTune()
-  hpt.report_hyperparameter_tuning_metric(
-      hyperparameter_metric_tag="answer",
-      metric_value=42)
+  Args:
+    data: the image to be written
+    path_prefix: the path prefix to write to
+
+  Returns:
+    {"min": value, "max": value} if rescaling was applied, None otherwise.
+  """
+  assert len(data.shape) == 4, data.shape
+  scaling = None
+  if data.dtype in [np.uint32, np.uint64]:
+    max_value = np.amax(data)
+    if max_value > 65535:
+      logger.warning("max_value %d exceeds uint16 bounds for %s.",
+                     max_value, path_prefix)
+    data = data.astype(np.uint16)
+  elif data.dtype in [np.float32, np.float64]:
+    min_value = np.amin(data)
+    max_value = np.amax(data)
+    scaling = {"min": min_value.item(), "max": max_value.item()}
+    data = (data - min_value) * 65535 / (max_value - min_value)
+    data = data.astype(np.uint16)
+  elif data.dtype in [np.uint8, np.uint16]:
+    pass
+  else:
+    raise NotImplementedError(f"Cannot handle {data.dtype}.")
+  bitdepth = 8 if data.dtype == np.uint8 else 16
+  greyscale = (data.shape[-1] == 1)
+  alpha = (data.shape[-1] == 4)
+  w = png.Writer(data.shape[1], data.shape[2], greyscale=greyscale,
+                 bitdepth=bitdepth, alpha=alpha)
+  for i in range(data.shape[0]):
+    img = data[i].copy()
+    if img.shape[-1] == 2:
+      # Pad two-channel images with a zero channel.
+      img = np.concatenate([img, np.zeros_like(img[..., :1])], axis=-1)
+    # pypng expects 2d arrays
+    # see https://pypng.readthedocs.io/en/latest/ex.html#reshaping
+    img = img.reshape(img.shape[0], -1)
+    with tf.io.gfile.GFile(f"{path_prefix}_{i:05d}.png", "wb") as fp:
+      w.write(fp, img)
+  return scaling
+
+
+def write_tiff(
+    all_imgs: np.ndarray,
+    path_prefix: tfds.core.ReadWritePath,
+) -> None:
+  """Save single-channel float images as tiff.."""
+  assert len(all_imgs.shape) == 4
+  assert all_imgs.shape[-1] == 1  # Single channel image
+  assert all_imgs.dtype in [np.float32, np.float64]
+  for i, img in enumerate(all_imgs):
+    path = path_prefix.parent / f"{path_prefix.name}_{i:05d}.tiff"
+
+    # Save tiff in an intermediate buffer
+    buffer = io.BytesIO()
+    imageio.imwrite(buffer, img, format=".tif")
+    path.write_bytes(buffer.getvalue())
+
+
+def write_single_record(
+    kv: Tuple[str, np.array],
+    path_prefix: PathLike,
+    scalings: Dict[str, Dict[str, float]],
+    lock: threading.Lock,
+):
+  """Write single record."""
+  key = kv[0]
+  img = kv[1]
+  if key == "depth":
+    write_tiff(img, path_prefix / key)
+  else:
+    scaling = write_scaled_png(img, path_prefix / key)
+    if scaling:
+      with lock:
+        scalings[key] = scaling
+
+
+def write_image_dict(data: Dict[str, np.array], path_prefix: PathLike):
+  # Pre-load image libs to avoid race-condition in multi-thread.
+  imageio.plugins.tifffile.load_lib()
+
+  scalings = {}
+
+  lock = threading.Lock()
+  _MAX_WRITE_THREADS = 16
+  with multiprocessing.pool.ThreadPool(
+      min(len(data.items()), _MAX_WRITE_THREADS)) as pool:
+    args = [(key, img) for key, img in data.items()]
+    write_single_record_fn = functools.partial(
+        write_single_record,
+        path_prefix=path_prefix,
+        scalings=scalings,
+        lock=lock)
+    pool.map(write_single_record_fn, args)
+    pool.close()
+    pool.join()
+
+  with tf.io.gfile.GFile(path_prefix / "data_ranges.json", "w") as fp:
+    json.dump(scalings, fp)
+
+
+def read_png(path: PathLike):
+  path = tfds.core.as_path(path)
+  pngReader = png.Reader(bytes=path.read_bytes())
+  width, height, pngdata, info = pngReader.read()
+  del pngReader
+  bitdepth = info["bitdepth"]
+  if bitdepth == 8:
+    dtype = np.uint8
+  elif bitdepth == 16:
+    dtype = np.uint16
+  else:
+    raise NotImplementedError(f"Unsupported bitdepth: {bitdepth}")
+  plane_count = info["planes"]
+  pngdata = np.vstack(list(map(dtype, pngdata)))
+  return pngdata.reshape((height, width, plane_count))
